Move the console code to a library so it can be used without the binary
This is related to https://github.com/gobuffalo/buffalo/issues/287
