+++ console/node_test.go
@@ -1,4 +1,4 @@
-package main
+package console
 
 import (
 	"go/parser"
