+++ console/quickfix.go
@@ -1,4 +1,4 @@
-package main
+package console
 
 import (
 	"strings"
