+++ console/terminal_unix.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package main
+package console
 
 import (
 	"fmt"
